+++ t/apicast-policy-logging.t
@@ -585,3 +585,61 @@ OK
 [qr/^Status\:\:200 USER_KEY\:\:123 42/]
 --- no_error_log
 [error]
+
+=== TEST 14: APICAST_ACCESS_LOG_BUFFER env parameter
+When buffer is enabled, log will be bump in chunks
+--- env eval
+('APICAST_ACCESS_LOG_BUFFER' => '1k')
+--- configuration
+{
+  "services": [
+    {
+      "id": 42,
+      "backend_version":  1,
+      "backend_authentication_type": "service_token",
+      "backend_authentication_value": "token-value",
+      "proxy": {
+        "hosts": [
+          "localhost"
+        ],
+        "api_backend": "http://test:$TEST_NGINX_SERVER_PORT/",
+        "proxy_rules": [
+          { "pattern": "/", "http_method": "GET", "metric_system_name": "hits", "delta": 1 }
+        ],
+        "policy_chain": [
+          {
+            "name": "apicast.policy.logging",
+            "configuration": {
+              "custom_logging": "Status::{{ status }} USER_KEY::{{credentials.user_key }} {{service.id}}",
+              "enable_json_logs": false
+            }
+          },
+          {
+            "name": "apicast",
+            "version": "builtin",
+            "configuration": {}
+          }
+        ]
+      }
+    }
+  ]
+}
+--- backend
+location /transactions/authrep.xml {
+  content_by_lua_block {
+    ngx.exit(200)
+  }
+}
+--- upstream env
+location / {
+  access_by_lua_block {
+      ngx.say("OK")
+  }
+}
+--- request
+GET /?user_key=123
+--- response_body env
+OK
+--- error_code: 200
+--- no_error_log eval
+[qr/^Status\:\:200 USER_KEY\:\:123 42/]